requirements.txt
cryptography
numpy
pillow
requests